Moved Checkbox and Radio button to Element folder and updated test for deleting project
