+++ TMS_Tests/appsettings.json
@@ -2,7 +2,7 @@
   "BrowserType": "chrome",
   "TimeOut": 10,
   "BaseUrl": "https://www.saucedemo.com/",
-  "BaseTRUrl": "https://qac0402.testrail.io/",
+  "BaseTRUrl": "https://qac0404.testrail.io/",
   "UserNameSauceDemo": "standard_user",
   "PasswordSauceDemo": "secret_sauce",
   "TRlogin": "workandreystep@gmail.com",
